Move route builder class and change import paths.
